requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==5.0.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.110.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.3
celery[redis]==5.3.6
redis==5.0.1
aiosmtplib==3.0.1
SQLAlchemy==2.0.23
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
langchain-openai>=0.1.0
openai>=1.0.0